% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.R
\name{Game}
\alias{Game}
\title{Guess Game!}
\usage{
Game(range = 10, trials = 3)
}
\arguments{
\item{range}{Range of generated inter}

\item{trials}{Number of your trials}
}
\description{
You can now play your favourite game! There is randomly generated integer in range [1,range]. You have some number of trials. Game will tell if your guess is greater or smaller than the hidden number.
}
